change board from api response(not instantly refreshed)
